Hide posts and edit buttons for unauthorized users

+++ resources/views/admin/posts/index.blade.php
@@ -15,6 +15,7 @@
                     </div>
 
                     <div class="panel-body">
+
                         <table class="table">
                             <thead>
                                 <tr>
@@ -48,8 +49,10 @@
                                                 <a href="{{ url("/admin/posts/{$post->id}/publish") }}" data-method="PUT" data-token="{{ csrf_token() }}" data-confirm="Are you sure?" class="btn btn-xs btn-warning">{{ $label }}</a>
                                             @endif
                                             <a href="{{ url("/admin/posts/{$post->id}") }}" class="btn btn-xs btn-success">Show</a>
-                                            <a href="{{ url("/admin/posts/{$post->id}/edit") }}" class="btn btn-xs btn-info">Edit</a>
-                                            <a href="{{ url("/admin/posts/{$post->id}") }}" data-method="DELETE" data-token="{{ csrf_token() }}" data-confirm="Are you sure?" class="btn btn-xs btn-danger">Delete</a>
+                                            @if (Auth::user()->is_admin || $post->user->id === Auth::user()->id)
+                                                <a href="{{ url("/admin/posts/{$post->id}/edit") }}" class="btn btn-xs btn-info">Edit</a>
+                                                <a href="{{ url("/admin/posts/{$post->id}") }}" data-method="DELETE" data-token="{{ csrf_token() }}" data-confirm="Are you sure?" class="btn btn-xs btn-danger">Delete</a>
+                                            @endif
                                         </td>
                                     </tr>
                                 @empty
